Issue 444: Improve postgraphile performance (#70)
* add webpack to deployment process

* add script to get Stack outputs

* add script to check Migrations

* fix getStackOutputs.js bug

* fix bugs in deploy.sh

* re-add sourcing of seed_flag

* fix output name in webpack config

* add extra logs for debugging

* fix postgraphile.cache zipping

* remove console logs

* use correct sls package when redeploying graphql Handler

* update graphql lambda version on re-deployment

* use express pattern for /graphql endpoint

* use postgraphile locally without open event listeners

* clean up local postgraphile endpoint

* fix hanging pg connections after tests run

* clean up tests broken during rebase

* fix db pool closing

* add NODE_ENV to dev deployments

* install dev dependencies during build step, even if NODE_ENV=production

* use lambda-proxy integration

* add test logs

* add migrations logs

* fix typo in checkMigrations script

* remove logs

* test logs for db connection

* try adding explicit DB URL

* try floodsPool without any options

* try plugging in object to postgraphile

* use pool object config for graphqlHandler

* add cors middleware

+++ deployment/deploy.sh
@@ -34,7 +34,7 @@ fi
 
 # Install "serverless" module and plugins
 yarn global add serverless@1.32.0
-yarn
+yarn --production=false
 
 # Create s3 bucket
 node $CURRENT_DIR/createS3Bucket.js
@@ -43,14 +43,58 @@ if [ $? != 0 ]; then
   exit 1
 fi
 
-# Deploy with serverless
-sls deploy -v | tee out.tmp
-if [ "${PIPESTATUS[0]}" != "0" ]; then
+# Get Stack output variables if Stack already exists
+node $CURRENT_DIR/getStackOutputs.js
+if [ $? != 0 ]; then
+  echo "db flag script failed"
+  exit 1
+fi
+source $CURRENT_DIR/stack_outputs.tmp
+
+if [ ! -z $PG_ENDPOINT ]; then
+  # If Stack and Database exist, check if our migrations are up to date
+  node $CURRENT_DIR/checkMigrations.js
+  if [ $? != 0 ]; then
+    echo "migration check failed"
+    exit 1
+  fi
+  source $CURRENT_DIR/migrations_flag.tmp
+else
+  MIGRATIONS_UP_TO_DATE=false
+fi
+
+# Create Serverless Bundle
+sls package -v
+if [ $? != 0 ]; then
+  echo "sls package failed"
+  exit 1
+fi
+
+# Add postgraphile.cache to bundle if deployed migrations are up to date
+if [[ $MIGRATIONS_UP_TO_DATE = "true" ]]; then
+  bash $CURRENT_DIR/bundleGraphqlHandler.sh
+  if [ $? != 0 ]; then
+    exit 1
+  fi
+fi
+
+# Deploy bundled serverless package
+sls deploy -p $CURRENT_DIR/../.serverless -v
+if [ $? != 0 ]; then
   echo "sls deploy failed"
   exit 1
 fi
-export PG_ENDPOINT=$(grep "PgEndpoint" out.tmp | cut -f2- -d: | cut -c2-)
-export GRAPHQL_ENDPOINT=$(grep "GraphqlEndpoint" out.tmp | cut -f2- -d: | cut -c2-)
+
+# Source variables if Stack is new
+if [[ $STACK_EXISTS = "false" ]]; then
+  echo "Inside stricter STACK_EXISTS clause"
+  node $CURRENT_DIR/getStackOutputsStrict.js
+  if [ $? != 0 ]; then
+    echo "stack output sourcing failed"
+    exit 1
+  fi
+  source $CURRENT_DIR/stack_outputs.tmp
+fi
 
 # Check if seeding will be required
 node $CURRENT_DIR/writeSeedFlag.js
@@ -73,9 +117,9 @@ if [ $? != 0 ]; then
   exit 1
 fi
 
-# If its a new deployment (as indicated by createS3Bucket.js), then seed data
+# If its a new deployment (as indicated by writeSeedFlag.js), then seed data
 source $CURRENT_DIR/seed_flag.tmp
-if [ $SEED_FLAG = "true" ]; then
+if [[ $SEED_FLAG = "true" ]]; then
   node $CURRENT_DIR/../db/scripts/seed.js
   if [ $? != 0 ]; then
     echo "seed script failed"
@@ -83,20 +127,30 @@ if [ $SEED_FLAG = "true" ]; then
   fi
 fi
 
-# Build Graphql Schema
-echo Building Schema
-node $CURRENT_DIR/../pgCatalog/buildPgCatalog.js
-if [ $? != 0 ]; then
-  echo "buildPgCatalog failed"
-  exit 1
+# If CloudFormation Stack is new or if migrations were not up to date, then re-deploy graphql Lambda function with postgraphile schema
+if [[ $STACK_EXISTS = "false" ]] || [[ $MIGRATIONS_UP_TO_DATE = "false" ]]; then
+  sls package -v
+  if [ $? != 0 ]; then
+    echo "sls package failed"
+    exit 1
+  fi
+
+  bash $CURRENT_DIR/bundleGraphqlHandler.sh
+  if [ $? != 0 ]; then
+    exit 1
+  fi
+
+  # Note: "sls -f graphql" is incompatible with "sls -p ...", so we must redeploy everything during first deployment
+  sls deploy -p $CURRENT_DIR/../.serverless
+  if [ $? != 0 ]; then
+    echo "sls deploy (with new graphql bundle) failed"
+    exit 1
+  fi
 fi
 
-# Deploy graphql Lambda function with schema
-sls deploy -f graphql
-if [ $? != 0 ]; then
-  echo "sls deploy -f graphql failed"
-  exit 1
-fi
-
-rm $CURRENT_DIR/seed_flag.tmp
-rm out.tmp
+# Clean up .tmp files
+[ -f $CURRENT_DIR/push_notification_flag.tmp ] && rm $CURRENT_DIR/push_notification_flag.tmp
+[ -f $CURRENT_DIR/custom_aws_service_name.tmp ] && rm $CURRENT_DIR/custom_aws_service_name.tmp
+[ -f $CURRENT_DIR/seed_flag.tmp ] && rm $CURRENT_DIR/seed_flag.tmp
+[ -f $CURRENT_DIR/stack_outputs.tmp ] && rm $CURRENT_DIR/stack_outputs.tmp
+[ -f $CURRENT_DIR/migrations_flag.tmp ] && rm $CURRENT_DIR/migrations_flag.tmp
